documentation: Add CONTRIBUTING.md
Providing a guideline for contributing can help with maintaining consistency throughout the repo. I added a CONTRIBUTING.md to further promote that.

Because of this, some things in the source code may need to be changed to satisfy this, but obviously not our current commit messages because that would take too long.