Fix: Grammar Ambiguity w/ Command Chain
This fixes ambiguity which arised within function call expressions
along with ones within addition. Command chains viewed `1 + 1` as:

    1(+1)

which caused this issue. Additionally the `ArgumentCallHead` had
unusual construction so I've removed that (or at least disabled) so
we have a working parser now. Hopefully this will allow all the
tests and all to pass and we can continue dev.

Signed-off-by: Vihan B <contact@vihan.org>