grammar: Merge all grammars into one "megaparser"
Like the name? It's only 431 lines anyways so you don't have to keep a
bunch of tabs open.

This also gets rid of a nasty circular dependency between parser.ne and
function.ne because both need a production rule from eachother. An ugly
hack to resolve this was by using a template rule.

ImmutableDictionary was removed because pretty useless anyways,
considering the fact that the production rule is already commented out.
I'm not sure what was being thought there with adding an "immutable"
dictionary to begin with.

The 4 literals Array, Dictionary, Tuple, and Set had Node constructors
that take in some VSLTokenType thing which isn't really needed and it
broke those current production rules, so the type argument was removed
from their Node constructors.

CommandChains took up a bunch of space in the grammar and, since it's
commented out, it should be removed. We have version control anyways so
we can restore it if we really need it.

A short PatternDeclaration production rule had the same fate here, and
TypeAlias wasn't referenced anywhere else so that was removed, too.

Lastly, some production rules were rearranged to make the grammar
easier to read.