feature: Improve parser (#30)
Firstly, all the grammar files were merged into one big "megaparser".
Like the name? It's only 459 lines anyways so you don't have to keep a
bunch of tabs open.

This also gets rid of a nasty circular dependency between parser.ne and
function.ne because both need a production rule from each other. An ugly
hack to resolve this was by using a template rule.

ImmutableDictionary was removed because it's pretty useless anyways,
considering the fact that the production rule is already commented out.
It may be re-added later, though, thanks to version control.

The 4 literals Array, Dictionary, Tuple, and Set had Node constructors
that take in some VSLTokenType thing which isn't really needed and it
broke those current production rules, so the type argument was removed
from their Node constructors.

CommandChains took up a bunch of space in the grammar and, since it's
commented out, it was removed. We have version control anyways so we can
restore it if we really need it.

A short PatternDeclaration production rule had the same fate here, and
TypeAlias wasn't referenced anywhere else so that was removed, too.

Some production rules were rearranged to make the grammar easier to
read, and some section comments were added to give this huge file a
sense of structure.

That grammar file really needs to get out every once in a while so it
moved out of its now-lonely grammar folder and out into the parser
folder.

Allowing no arguments in a function call fixes #31 with a minor change
to the grammar. Lastly, some tests were added for that fix.

Signed-off-by: CrazyGuy108 <taylor@hansenfamily.us>