feature: revamp CLI code (#32)
CLI previously was in one file, not very extensible and clunky. This
splits the CLI into an OO, multi-file system which follows OO prinicple
and is simple to extend and add flags. This changes flags but adds many
features including:

 - Property TTY deduction
 - Smart color disable/enabling
 - STDIN v File support
 - Smarter REPL

Additionally this supports subcommands which have not beed added ye
except for `vsl run` which is an alias for `vsl`

Signed-off-by: Vihan B <contact@vihan.org>