feature: registers functions and clean up (#29)
* feature: registers functions and clean up

This does a lot of things including fixing many many bugs along with
adding function registration. Lookup is not added yet.

* Major work on resolution of functions

Function resolution is arguably one of the more complex parts of VSL so
making good progress on that. Currently is using RL filling but if
possible, a better algorithm should be implemented in the future.

* Major work on resolution of functions

Function resolution is arguably one of the more complex parts of VSL so
making good progress on that. Currently is using RL filling but if
possible, a better algorithm should be implemented in the future.

* grammar: Add constructors (#35)

Signed-off-by: CrazyGuy108 <taylor@hansenfamily.us>

* Fixed recursive porperty bork resulting in bork.

Previously recursive porerpty would bork beacuse
optional but also not optional and like wat is
hapen wat because not usposed to bork but is
actually bork because position so like wat hapen
and lot confusing

* Modified debug flags for -d prefix

* internal -> external

Makes like zero sense to be called internal because it specifies
external linkage anyway so just fixed that

* Context & scope persistance across repl lines.

This adds basic support for context propogation through the REPL so I
can declare a class in one line and use it later.

Signed-off-by: Vihan B <contact@vihan.org>