feature: add plane intersection resolution (#38)
This adds plane intersection resolution. This adds deduction using two sets of
types. Since we're trying to 'fit' an intersection into the root and applied,
you can think of the types as a three-dimentional figure which we're trying to
find an _intersection plane_ which satisfies our problem. This uses the three
types:

  - Root: A type we're trying to fit an applied type into
  - Applied: A type that should fit into at least one root
  - Intersection: A tuple of a (Root, Applied) pair which satisfies
                  each other. We can call this tuple a 'derived candidate'

Signed-off-by: Vihan B <contact@vihan.org>