feature: add FIX-ITs (#48)
This adds 'FIX-IT's which automatically apply suggestions to solve certain
errors. At the moment pretty experimental due to bad data from the tokenizer but
the FIX-ITs itself work. This seperates into three main parts:

 - FixIt Controller
 - FixIt Agent
 - FixIt Manager
 - FixIt

The controller is a MVC component which allows for plugging in FIX-ITs to a
basic GUI textual format. The agent itself applies a FixIt to a code. The
manager is what does the actual fixing, it is managed by a FixItAgent instance
which takes a FixIt instance from an object and applies it.

Signed-off-by: Vihan B <contact@vihan.org>