merge: add command chain and a ton of other thnigs (#61)
This adds a lot of things from C9, command chains, recursive types, tokenizer upgrades, block statements, and a lot more.

---

* test: Added regenerator tests.

This adds basic regenerator tests for expressions. The regenerator
is the thing which takes an AST and is supposed to output a valid
string which should evaluate to the same thing. It can also be
used to inspect the AST to ensure precedence and the actual AST is
correct.

* fix: made `**` right associative.
* parser: Add mostly unambiguous  command chains to grammar
* feature: Recursive stuff and command chains
* fix: Change let to const and var to let, add tests for stdlib and examples, add to grammar
* fix: typealias errors
* fix: error position data
* feature: AST viewer using -asts
* fix: lot of bug